license_server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
pydantic==2.9.2
